+++ src/worldbuilder.h
@@ -19,7 +19,7 @@ class WorldBuilder : public osg::Referenced {
 			m_maxBuildingsPerFile(100) {};
 		void setConfigFilename(std::string configFilename) { m_configFilename = configFilename; }
 		void setOutputFilename(std::string outputFilename) { m_outputFilename = outputFilename; }
-		void setMaximumBuildingsPerFile(size_t number) { m_maxBuildingsPerFile = number; }
+		void setMaximumBuildingsPerFile(int number) { m_maxBuildingsPerFile = number; }
 		void loadConfiguration();
 		void buildWorld();
 	protected:
@@ -30,7 +30,7 @@ class WorldBuilder : public osg::Referenced {
 		std::string m_configFilename;
 		std::string m_outputFilename;
 
-		size_t m_maxBuildingsPerFile;
+		int m_maxBuildingsPerFile;
 
 		std::vector<std::string> m_shapeFilenames;
 		std::vector<std::string> m_heigthFilenames;
